apply repository and service to PIM.Web

+++ PIM.Common/Constants/GlobalConfigurationConstants.cs
@@ -3,6 +3,6 @@
     public static class GlobalConfigurationConstants
     {
         public const string CONNECT_STRING_NAME = "ProjectExcerciseContext";
-        public const short ADONET_BATCH_SIZE = 20;
+        public const int ADONET_BATCH_SIZE = 20;
     }
 }
